% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_time_offset.R
\name{calc_time_offset}
\alias{calc_time_offset}
\title{Calculate UTC time offset for a location and set of dates}
\usage{
calc_time_offset(dates, lat, lon)
}
\arguments{
\item{dates}{POSIXct vector. Date-times to calculate offsets for.}

\item{lat}{Numeric. Latitude in decimal degrees.}

\item{lon}{Numeric. Longitude in decimal degrees.}
}
\value{
Numeric vector of UTC offsets in hours (e.g. 1 for CET, 2 for CEST).
}
\description{
Looks up the time zone for a given coordinate and returns the UTC offset
in hours for each date-time, accounting for daylight saving transitions.
}
\examples{
dates = as.POSIXct(c("2024-06-15 12:00", "2024-12-15 12:00"), tz = "UTC")
calc_time_offset(dates, lat = 52.11, lon = 5.03)

}
