% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gb_lidar.R
\name{get_gb_lidar}
\alias{get_gb_lidar}
\title{Download GB LiDAR elevation data from Environment Agency}
\usage{
get_gb_lidar(
  domain,
  product = c("fz_dsm", "dsm", "dtm", "vom"),
  product_type = c("elevation", "hillshade"),
  res = c(1, 2)
)
}
\arguments{
\item{domain}{An \code{sf} object defining the area of interest (any CRS;
will be transformed to EPSG:27700 internally).}

\item{product}{Character. Either \code{"DSM"} or \code{"DTM"}.}

\item{product_type}{Character. Either \code{"elevation"} or \code{"hillshade"}.
Currently unused but reserved for future support.}

\item{res}{Numeric. Resolution in metres (1 or 2). Currently unused;
resolution is determined by the EA composite dataset.}
}
\value{
A \code{SpatRaster} of the requested elevation product.
}
\description{
Retrieves DSM or DTM raster data from the Environment Agency's LiDAR
composite catalogue for England via the \code{gblidar} package.
}
\examples{
\dontrun{
domain <- get_custom_domain(latitude = 51.45, longitude = -2.59)
dsm <- get_gb_lidar(domain, product = "DSM")
dtm <- get_gb_lidar(domain, product = "DTM")
}

}
