% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nl_lidar.R
\name{get_nl_lidar}
\alias{get_nl_lidar}
\title{Download Dutch AHN LiDAR elevation data}
\usage{
get_nl_lidar(domain, product = c("DSM", "DTM"), res = c(0.5, 5))
}
\arguments{
\item{domain}{An \code{sf} object defining the area of interest (any CRS;
will be transformed to EPSG:28992 internally).}

\item{product}{Character. Either \code{"DSM"} or \code{"DTM"}.}

\item{res}{Numeric. Raster resolution in metres. AHN supports 0.5 m and 5 m.}
}
\value{
A \code{SpatRaster} (or raster object, depending on \code{rAHNextract}
version) of the requested elevation product.
}
\description{
Retrieves DSM or DTM raster data from the Dutch AHN (Actueel Hoogtebestand
Nederland) Web Coverage Service for a given study area.
}
\examples{
\dontrun{
domain <- get_custom_domain()
dsm <- get_nl_lidar(domain, product = "DSM", res = 0.5)
dtm <- get_nl_lidar(domain, product = "DTM", res = 5)
}

}
